Create FailedAttempt
This is what we were working on to try to generate enemies, generate carrot bullets to hit the enemies, and deal with collisions between the bullets and the enemies, and the enemies and the avatar. We kept encountering errors and couldn't figure out how to fix them, especially the errors relating to the ipairs for bullets and enemies.